+++ public/components/community/package.go
@@ -31,6 +31,7 @@ import (
 	_ "github.com/redpanda-data/connect/v4/public/components/confluent"
 	_ "github.com/redpanda-data/connect/v4/public/components/couchbase"
 	_ "github.com/redpanda-data/connect/v4/public/components/crypto"
+	_ "github.com/redpanda-data/connect/v4/public/components/cypher"
 	_ "github.com/redpanda-data/connect/v4/public/components/dgraph"
 	_ "github.com/redpanda-data/connect/v4/public/components/discord"
 	_ "github.com/redpanda-data/connect/v4/public/components/elasticsearch"